compatibility to DynaMind Core v0.4
the API to access cell values has change in the new DynaMind-Core v0.4 getValue returns now the value of the actual coordinates. For the old functionality use getCell.

To set the version the parameter self.DynaMind_Core_Version has been added in the constructor. If the parameter is set to 0.4 the new API is used.